+++ client/src/test/integration/GroupPage.spec.tsx
@@ -28,7 +28,7 @@ describe('Group page', () => {
       const { findAllByText } = render(<AppRender route={page} />)
 
       await findAllByText(/1990/i)
-      const date = screen.getByText('4/10/1990')
+      const date = screen.getByText('4/10/1990 CE')
       expect(date).toBeInTheDocument()
     })
 
@@ -44,7 +44,7 @@ describe('Group page', () => {
       const { findAllByText } = render(<AppRender route={page} />)
 
       await findAllByText(/1994/i)
-      const date = screen.getByText('5/13/1994')
+      const date = screen.getByText('5/13/1994 CE')
       expect(date).toBeInTheDocument()
     })