+++ client/CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Changed parsing of person dates ([#310](https://github.com/project-lux/lux-frontend/issues/310)).
+
 ### Removed
 
 ### Fixed